+ | Added service executing funcs (not tested yet)
