+++ internal/services/services.go
@@ -1,63 +1,44 @@
-package main
+package services
 
 import (
-	"fmt"
-	"log"
 	"os"
-	"github.com/alecthomas/participle/v2"
-	"github.com/alecthomas/participle/v2/lexer"
+	"exec"
+	"strconv"
+	"strings"
+	"github.com/demostanis/userspace_digressions/internal/initctl"
 )
 
-type Service struct {
-	Entries []*Entry `@@*`
-}
-
-type Entry struct {
-	Key		string `@Ident "="`
-	Value	string `@Ident`
-}
-
-var (
-	serviceLexer = lexer.MustSimple([]lexer.SimpleRule{
-		{"Ident", `[a-zA-Z0-9/_\-.:,]+`},
-		{"Punct", `=`},
-		{"whitespace", `\s+`},
-	})
-	parser = participle.MustBuild[Service](
-		participle.Lexer(serviceLexer),
-	)
+const (
+	SERVICE = 0
+	COMMAND = 1
+	RUNLEVEL = 2
 )
 
-func CheckServiceValidity(service *Service) error {
-	if service.Entries[0].Key != "Service" {
-		return fmt.Errorf("line 1: no service specified")
-	}
-	if service.Entries[1].Key != "Command" {
-		return fmt.Errorf("line 2: no command specified")
-	}
-	if service.Entries[2].Key != "RunLevel" {
-		return fmt.Errorf("line 3: no run level specified")
+func ExecCommand(command string) error {
+	args := strings.Split(command)
+	err := exec.Command("bash", args...)
+	return err
+}
+
+func runServices(services []Service, runLevel int) error {
+	var errors []string
+	for _, el := range services {
+		elLevel, err :=  strconv.Atoi(el.Entries[RUNLEVEL].Value)
+		if err != nil {
+			errors = append(errors, fmt.Sprintf("failed to load service %s: %v", el.Entries[SERVICE].Value));
+			continue
+		}
+		if elLevel == runLevel {
+			err = ExecCommand(el.Entries[COMMAND].Value)
+			if err != nil {
+				errors = append(errors, fmt.Sprintf("failed to load service %s: %v", el.Entries[SERVICE].Value));
+				continue
+			}
+		}
 	}
 
+	if len(errors) > 0 {
+		return fmt.Errorf("encountered errors:\n%s", strings.Join(errors, "\n"))
+	}
 	return nil
-}
-
-func ServiceParser(fileName string) (Service, error) {
-	file, err := os.ReadFile(fileName)
-	if err != nil {
-		return Service{}, fmt.Errorf("error opening file: %w", err)
-	}
-	content := string(file)
-
-	service, err := parser.ParseString("", content)
-	if err != nil {
-		return Service{},fmt.Errorf("error parsing file: %w", err)
-	}
-
-	err = CheckServiceValidity(service)
-	if err != nil {
-		return Service{},fmt.Errorf("error parsing file: %w", err)
-	}
-
-	return *service, nil
 }
